Added ability to undo more than one tile at a time
Initially the undo stack recorded one tile change at a time. Now the
undo stack records all changes while the mouse button is held down as
one object. This means that when you click and drag to draw a lot of
tiles at once, when you press undo, it will undo all of the tiles that
were drawn and not just one at a time.