Start refactoring of TileMap and related classes
TileMap has been using a 2D vector of tile objects. But because tiles
are immutable and the same regardless of where they are placed, it is a
waste of memory to have a copies of the same object. This refactor will
instead use the flyweight pattern and have one Tile object per type. The
tile map will then just have a 2D vector of tile pointers pointing to
their respective type. This touches a lot of different pieces of
code,including the level editor and gameplay scene.