Add target_grid_index to game objs
Before there only use to be a grid_coord and thats it. This made it so
the object was considered to be at the target index before even actually
reaching. This had some issues, including the bears in a way
'predicting' where you'll end up on the ice. Now the object is on it's
grid_index, and only changes to its target index once it truly reaches
it