Change BFS to use 2D array of Vec2
Because I have yet to implement a hashing function for Vec2, BFS was
using a string to Vec2 mapping. This made it so that every time I used
the seen/parent map, I had to convert the Vec2 to a string. The
percieved benefit of the map over the 2D array was saved memory, because
a 2D array would have over 600 tiles in it. However, the map size was
regularly 512 buckets, making the savings minimal.