Remove Direction enum and replace with dir vector
The Direction enum was only used during the GetSlidingTargetTile func
and to get a vector in which to get the next tile. Because of this it
makes more sense to just use an Vec2 direction directly, instead of
using the enum as an intermediate step. This also results in the sliding
target tile function being a lot more concise.