Change how fish are spawned
Previously fish would just be spawned on any ice tile that exists within
the map. This lead to the map looking very busy and messy, especially
with an actual fish texture. Now, fish must be placed on the level.txt
file explicitly. This cleans up the look of the map and allows for
strategic placement of fish to make the level as challenging as you
want.