Fix: Update LevelEditor controls and reset button
Using double click to place ground tile resulted in a different tile
being placed on the first click before placing the ground. This just
looked weird. I have changed the controls to be l-click for ice, r-click
for walls. 'F' for fish and 'G' for ground.

Also fixed reset button not getting rid of fish or moving the objects
back to starting point