Fix: Fix sliding pathfinding
Sliding pathfinding had the issue of only being able to find tiles that
it could actually land on(i.e ground/corner ice tiles). This caused
issues for Bjorn pathfinding if Bruce's starting position was off. It
also caused issues with checking to see if Bruce could reach all fish in
the level editor. This fix changes the BFS function to check all tiles
that are slid through to make sure the endpoint is not passed.