Fix: Fix issues when creating/deleting levels
Fix crash that occured when deleting a level. This was due to the level
value being incremented in order to rename all previous custom levels.
Changing the order of operations fixed this.

Also fixed bug where saving a level would increment the
num_custom_levels value, which resulted in unusable LevelCards being
created.

- Also removed is_custom parameter from LevelEditorScene's constructor,
  because it's always true