Fix: Fix crash when selecting non-custom level
This crash occured because, the level select lambda for the play button
takes a reference to the 'show_custom' parameter of the ctor, which goes
out of scope after the ctor finishes. So when I replace the scene, that
boolean value is undefined and therefor UB. Capturing that boolean by
value fixes the problem.